requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.10.0
webdriver-manager==3.8.6

# Natural Language Processing
nltk==3.8.1
spacy==3.6.1

# Vector database
faiss-cpu==1.7.4
chromadb==0.4.6

# API and web
fastapi==0.101.0
uvicorn==0.23.2
streamlit==1.25.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0